Add new hello world main file, rename old main file
