+++ Dockerfile
@@ -1,23 +1,22 @@
 FROM ghcr.io/ministryofjustice/analytical-platform-airflow-python-base:1.16.0@sha256:96551185eef8ff5f54f82b6a55c5df570a795d155851cde1811f230423efca1b
 
-# Below is an example of how to use the base image
 
 # Switch to root user to install packages
-# USER root                 
+USER root                 
                        
 # Copy requirements.txt
-# COPY requirements.txt requirements.txt 
+COPY requirements.txt requirements.txt 
 
 # Copy application code
-# COPY src/ .
+COPY src/ .
 
 # Install requirements
-# RUN <<EOF
-# pip install --no-cache-dir --requirement requirements.txt
-# EOF
+RUN <<EOF
+pip install --no-cache-dir --requirement requirements.txt
+EOF
 
 # Switch back to non-root user (analyticalplatform)
-# USER ${CONTAINER_UID}
+USER ${CONTAINER_UID}
 
 # Execute main.py script
-# ENTRYPOINT ["python3", "main.py"]
+ENTRYPOINT ["python3", "main.py"]
